refactor: ♻️ Redo get_devicegroups_from_pano to return dict.
So that the function can be used other places and commands I've moved the text message to be created in the chat command and the util function to just return a dictionary of data about the groups and devices in them. I've also updated the tests to account for this.